requirements.txt
asgiref==3.7.2
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
DateTime==5.2
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
gunicorn==21.2.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
model-bakery==1.17.0
outcome==1.2.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.3
pytest-django==4.6.0
pytest-testdox==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
webdriver-manager==4.0.1
whitenoise==6.6.0
wsproto==1.2.0
zope.interface==6.0
